Aded user secret
