--- neophyte_proj/DataAccess/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace DataAccess
-{
-    public class Class1
-    {
-
-    }
-}
