+++ neophyte_proj/DataAccess/Models/IntermediateModels/CourseTeacher.cs
@@ -1,7 +1,7 @@
-﻿using neophyte_proj.WebApi.Models.CourseModel;
-using neophyte_proj.WebApi.Models.TeacherModel;
+﻿using neophyte_proj.DataAccess.Models.CourseModel;
+using neophyte_proj.DataAccess.Models.TeacherModel;
 
-namespace neophyte_proj.WebApi.Models.IntermediateModels
+namespace neophyte_proj.DataAccess.Models.IntermediateModels
 {
     public class CourseTeacher
     {
