+++ neophyte_proj/DataAccess/Models/StudentModel/Student.cs
@@ -1,8 +1,8 @@
-﻿using neophyte_proj.WebApi.Models.CourseModel;
-using neophyte_proj.WebApi.Models.IntermediateModels;
-using neophyte_proj.WebApi.Models.TeacherModel;
+﻿using neophyte_proj.DataAccess.Models.CourseModel;
+using neophyte_proj.DataAccess.Models.IntermediateModels;
+using neophyte_proj.DataAccess.Models.TeacherModel;
 
-namespace neophyte_proj.WebApi.Models.StudentModel
+namespace neophyte_proj.DataAccess.Models.StudentModel
 {
     public class Student
     {
