+++ neophyte_proj/DataAccess/Models/StudentModel/StudentAccountInfo.cs
@@ -1,13 +1,15 @@
-﻿namespace neophyte_proj.WebApi.Models.TeacherModel
+﻿using System.Data;
+
+namespace neophyte_proj.DataAccess.Models.StudentModel
 {
-    public class TeacherAccountInfo
+    public class StudentAccountInfo
     {
         public int Id { get; set; }
         public string UserName { get; set; }
         public string Password { get; set; }
         public DateTime RegistrationDate { get; set; }
         //1to1
-        public int TeacherId { get; set; }
-        public Teacher Teacher { get; set; }
+        public int StudentId { get; set; }
+        public Student Student { get; set; }
     }
 }
